refactor js folder
