+++ src/platform/chrome/ts/browser-api.ts
