+++ src/ts/errors/authentication.error.ts
