+++ src/ts/helpers/context.helper.ts
@@ -1,9 +1,9 @@
-import {Service} from '@js/services/service';
-import {ApiService} from '@js/services/api.service';
-import {DavService} from '@js/services/dav.service';
-import {UserService} from '@js/services/user.service';
-import {StorageService} from '@js/services/storage.service';
-import {PromiseService} from '@js/services/promise.service';
+import {Service} from '@ts/services/service';
+import {ApiService} from '@ts/services/api.service';
+import {DavService} from '@ts/services/dav.service';
+import {UserService} from '@ts/services/user.service';
+import {StorageService} from '@ts/services/storage.service';
+import {PromiseService} from '@ts/services/promise.service';
 
 export class ContextHelper {
