+++ src/ts/options.ts
@@ -1,7 +1,8 @@
 import Vue from 'vue';
 import {MdButton} from 'vue-material/dist/components';
 import OptionsComponent from '@vue/options/OptionsComponent.vue'
-import {ContextHelper} from '@js/helpers/context.helper';
+import {ContextHelper} from '@ts/helpers/context.helper';
+import '@scss/main.scss';
 
 Vue.use(MdButton);
 ContextHelper.buildContext();
