+++ src/ts/services/promise.service.ts
@@ -1,4 +1,4 @@
-import {Service} from '@js/services/service';
+import {Service} from '@ts/services/service';
 
 export class PromiseService extends Service {
