--- src/vue/options/OptionsComponent.scss
@@ -1,1 +1,0 @@
-@import "~@scss/main.scss";
