--- src/vue/popup/PopupComponent.scss
@@ -1,1 +1,0 @@
-@import "~@scss/main.scss";
